加入私有函式設計

+++ src/ios/CLKPromises/CLKPromises/CPProgress.m
@@ -5,7 +5,7 @@
 
 
 // constructors
-- (instancetype) initWithLocalizedDescription:(NSString*)localizedDescription completedCount:(int)completedCount totalCount:(int)totalCount {
+- (instancetype) initWithLocalizedDescription:(NSString*)localizedDescription completedCount:(NSInteger)completedCount totalCount:(NSInteger)totalCount {
     
     // super
     self = [super init];
